chore(root): updated and migrated the Yarn version to v4
Q: Why not migrate to npm v10?
A: In the past, during the npm v7 days, I experienced various problems in
hoisting up. I tried this during the Yarn v4 transition, but I still have
this on hold due to difficulties with parallel execution of npm-scripts in
the workspace.

see: https://github.com/npm/rfcs/issues/610